[Backend] refactor: rename directory /ws to /chat
