+++ backend/internal/chat/domain/client.go
