+++ backend/internal/router/router.go
@@ -12,15 +12,15 @@ import (
 	email "github.com/unmsmfisi-socialapplication/social_app/internal/email_sender"
 	"github.com/unmsmfisi-socialapplication/social_app/internal/login/application"
 	"github.com/unmsmfisi-socialapplication/social_app/internal/login/infrastructure"
-	"github.com/unmsmfisi-socialapplication/social_app/internal/profile"
 	"github.com/unmsmfisi-socialapplication/social_app/internal/post"
+	"github.com/unmsmfisi-socialapplication/social_app/internal/profile"
 
 	interest_topics "github.com/unmsmfisi-socialapplication/social_app/internal/interest_topics"
 	registerapplication "github.com/unmsmfisi-socialapplication/social_app/internal/register/application"
 	registerinfrastructure "github.com/unmsmfisi-socialapplication/social_app/internal/register/infrastructure"
 	"github.com/unmsmfisi-socialapplication/social_app/pkg/database"
 
-	wsInf "github.com/unmsmfisi-socialapplication/social_app/internal/ws/infraestructure"
+	wsInf "github.com/unmsmfisi-socialapplication/social_app/internal/chat/infraestructure"
 
 	follow "github.com/unmsmfisi-socialapplication/social_app/internal/follow"
 )
@@ -43,7 +43,7 @@ func Router(wsHandler *wsInf.Handler) http.Handler {
 
 	postRoutes := post.PostModuleRouter(dbInstance)
 
-    profileRouter := profile.ProfileModuleRouter(dbInstance)
+	profileRouter := profile.ProfileModuleRouter(dbInstance)
 
 	r.Get("/", func(w http.ResponseWriter, r *http.Request) {
 		w.Write([]byte("{\"hello\": \"world\"}"))
@@ -61,7 +61,6 @@ func Router(wsHandler *wsInf.Handler) http.Handler {
 
 		w.Write([]byte(fmt.Sprintf("{\"response\": \"all done slow\"}")))
 	})
-
 
 	// Login
 	loginRepo := infrastructure.NewUserDBRepository(dbInstance)
